I changed ImageProcessor to have one unified method to return the newest spots. (I also took out that annoying canvas that we had popping up)
I implemented the ProcessingManager class. It basically calls ImageProcessor
to process an image on a timed loop. I made this class run on it's own thread
and it seems to work so far..... stay tuned 0_0

I changed the main class to use the processing manager instead of using the
processor directly.